Exit early if no movies - RIP Alamo Drafthouse Woodbury

+++ script.sh
@@ -7,10 +7,18 @@ CURRENT_LIST="current_movie_list.txt"
 OLD_LIST="previous_movie_list.txt"
 
 :> "$JSON_OUTPUT"
-curl "$URL" 2> /dev/null 1> "$JSON_OUTPUT"
+curl -31kL "$URL" 2> /dev/null 1> "$JSON_OUTPUT"
 
-if [ -s "$JSON_OUTPUT" ]; then
+if [ -s "$JSON_OUTPUT" ] && [ $(file "$JSON_OUTPUT" | grep -c 'HTML document') -eq 0 ]; then
+
     sed 's/"FilmName":/~/g' "$JSON_OUTPUT" | tr '~' '\n' | grep -v '{"Calendar":' | awk -F '"' '{print $2}' | sort | uniq > "$CURRENT_LIST"
+
+    # If no movies (RIP Woodbury Alamo Drafthouse -- July 2018 - May 2023)
+    if [ ! -s "$CURRENT_LIST" ]; then
+        :> "$JSON_OUTPUT"
+        rm "$CURRENT_LIST"
+        exit 1
+    fi
 
     if [ -s "$OLD_LIST" ]; then
         #Compare current to old movie list, find list of new movies.
@@ -19,8 +27,6 @@ if [ -s "$JSON_OUTPUT" ]; then
         if [ -n "$new_movies" ]; then
             echo "New movies at the alamo drafthouse:"
             echo "$new_movies"
-        else
-            echo "There are no new movies at the alamo drafthouse compared to last run."
         fi
         mv "$CURRENT_LIST" "$OLD_LIST"
     else
@@ -28,5 +34,8 @@ if [ -s "$JSON_OUTPUT" ]; then
         cat "$CURRENT_LIST"
         mv "$CURRENT_LIST" "$OLD_LIST"
     fi
-
+else
+    echo "ERROR IN GRABBING THE JSON OUTPUT FROM THE ALAMO CALENDAR API"
 fi
+
+:> "$JSON_OUTPUT"
